refactor(execution-mode): overhaul AppContext initialization and execution mode handling
- Add ExecutionMode enum and store it in AppContext
- Replace AppContext.getInstance() with initialize(mode) and expose getExecutionMode()
- Update CLI, desktop, and tests to use the new initialization
- Simplify ChartRenderer and move individual chart renderers to dedicated files
- Update README with new description, install script, and CLI link
- Increment project version to 1.2.5

BREAKING CHANGE: AppContext.initialize replaces the old getInstance and now requires an ExecutionMode argument; update all callers accordingly.
