+++ cli/src/test/kotlin/io/askimo/core/chat/service/ChatSessionServiceIT.kt
@@ -60,7 +60,7 @@ class ChatSessionServiceIT {
                 messageRepository = messageRepository,
                 sessionMemoryRepository = sessionMemoryRepository,
                 projectRepository = projectRepository,
-                appContext = AppContext.getInstance(),
+                appContext = AppContext.initialize(ExecutionMode.STATELESS_MODE),
             )
         }
