+++ gradle.properties
@@ -1,6 +1,6 @@
 # Project metadata
 projectGroup=io.askimo
-projectVersion=1.2.4
+projectVersion=1.2.5
 
 # About information
 author=Hai Nguyen
